refactor: central loader component
