+++ src/app/protected/[user]/page.tsx
@@ -6,22 +6,9 @@ import { useParams } from "next/navigation";
 import PostModal from "./(components)/PostModal";
 import Image from "next/image";
 import firstBlog from "@/assets/dazzle-blog-post-article.gif";
-import {
-  Card,
-  CardBody,
-  CardHeader,
-  CardFooter,
-  // Modal,
-  // ModalContent,
-  // ModalHeader,
-  // ModalBody,
-  // ModalFooter,
-  Button,
-  useDisclosure,
-} from "@nextui-org/react";
+import { Card, CardBody, CardHeader, CardFooter, Button, useDisclosure } from "@nextui-org/react";
 import { Blog } from "@/models/UserModel";
 import { FaEdit } from "react-icons/fa";
-import { MdDelete } from "react-icons/md";
 import Link from "next/link";
 import Loader from "@/app/signin/(components)/Loader";
 
@@ -63,7 +50,10 @@ export default function User() {
           <ul className="w-full flex flex-col items-center p-3">
             <h4 className="text-4xl mt-4  text-blue-700">Your Blogs</h4>
             {blogs.map((blog: Blog) => (
-              <li key={blog?._id} className="mt-5">
+              <li
+                key={blog?._id}
+                className="mt-5"
+              >
                 <Card className="w-[600px]  p-4 text-cyan-50 ">
                   <CardHeader className="text-2xl p-4">{blog.title}</CardHeader>
                   <CardBody className="p-4">{blog.content}</CardBody>
@@ -97,7 +87,12 @@ export default function User() {
       ) : (
         <div className="h-full flex flex-col justify-center items-center">
           {/* <h2 className="text-2xl"></h2> */}
-          <Image src={firstBlog} alt="first blog" height={400} width={400} />
+          <Image
+            src={firstBlog}
+            alt="first blog"
+            height={400}
+            width={400}
+          />
 
           <Button
             size="lg"
