+++ src/components/Loader.tsx
@@ -1,7 +1,4 @@
-"use client";
-
-import React from "react";
-import { CircularProgress } from "@nextui-org/react";
+import { CircularProgress } from "@nextui-org/progress";
 
 const Loader = () => {
   return (
