+++ src/components/Navbar.tsx
@@ -1,21 +1,16 @@
 "use client";
 
 import React from "react";
-import {
-  Navbar,
-  NavbarBrand,
-  NavbarContent,
-  NavbarItem,
-  Link,
-  Button,
-} from "@nextui-org/react";
+import { Navbar, NavbarBrand, NavbarContent, NavbarItem, Link, Button } from "@nextui-org/react";
 import { useSession, signOut } from "next-auth/react";
-import { FiEdit } from "react-icons/fi";
 
 const Appbar = () => {
   const { data: session } = useSession();
   return (
-    <Navbar className="bg-inherit h-[75px]" maxWidth="full">
+    <Navbar
+      className="bg-inherit h-[75px]"
+      maxWidth="full"
+    >
       <NavbarBrand>
         <Link href="/">
           <p className="font-bold text-inherit">blog-g-gers</p>
