+++ src/hooks/useFetchblogs.tsx
@@ -3,13 +3,11 @@
 import { useEffect, useState } from "react";
 
 import { Blog } from "@/models/UserModel";
-import { useBlogStore } from "@/utils/store";
 
 export const useFetchBlogs = () => {
   const [isLoading, setIsLoading] = useState<Boolean>();
 
   const [blog, setBlog] = useState<Blog[]>([]);
-  const [error, setError] = useState("");
   useEffect(() => {
     setIsLoading(true);
     const fetchData = async () => {
